% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNCrelated.R
\name{Check_hgnc_hits}
\alias{Check_hgnc_hits}
\title{修正HGNC的Multi-symbol checker工具的输出结果}
\usage{
Check_hgnc_hits(hgnc.hits)
}
\arguments{
\item{hgnc.hits}{HGNC在线工具'Multi-symbol checker'的结果, A data.frame}
}
\value{
A data.frame for UniqueHit or A list with two data.frame for UniqueHit and MultiHit
}
\description{
HGNC 提供的\href{https://www.genenames.org/tools/multi-symbol-checker/}{Multi-symbol checker在线工具}可以将基因名更新到最新,
但是会输出所有匹配的结果,也就是说单个input gene可能会有多个hits.此工具可保
留主要匹配结果,去除可能的错误hits以及未必对上的input.
}
