% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplePlot.R
\name{David_barplot}
\alias{David_barplot}
\title{David富集结果绘图 柱状图}
\usage{
David_barplot(
  df,
  term.prefix = c("(^GO:\\\\d+~)|(^\\\\w+\\\\d+[:~])"),
  fill.color = c("blue", "red"),
  show.top = 10,
  x = "Count",
  xlabel = "Gene Counts",
  arrange.by.x = FALSE,
  decresing = TRUE
)
}
\arguments{
\item{df}{David富集结果, A data.frame}

\item{term.prefix}{Term列中的前缀，会被去除}

\item{fill.color}{柱子的颜色,一个向量，分别对应最低值和最高值}

\item{show.top}{要展示的Term数目}

\item{x}{X轴的映射}

\item{xlabel}{X轴的标签}

\item{arrange.by.x}{Term是否按照X轴的值进行排序，默认是按照PValue排序的}

\item{decresing}{排序方式默认为降序}
}
\value{
a ggplot2 plot
}
\description{
David富集结果绘图 柱状图
}
\examples{
requireNamespace("dplyr")
requireNamespace("ggplot2")
df <- read.delim(system.file("extdata", "David_outputs_GO.txt",
package = "ZhangRtools"), stringsAsFactors = FALSE)
David_barplot(df, fill.color = c("#ff9999","#ff0000"),x = "Fold.Enrichment",
 xlabel = "Fold Enrichment")
David_barplot(df, x = "Fold.Enrichment", xlabel = "Fold Enrichment",
arrange.by.x = TRUE)
df \%>\% dplyr::mutate(fdr = -log(FDR, base=10)) \%>\% David_barplot(x = "fdr",
xlabel = "-log(10)FDR")
kegg.res <- read.delim(system.file("extdata", "David_outputs_KEGG.txt",
package = "ZhangRtools"), stringsAsFactors = FALSE)
David_barplot(df = kegg.res,  fill.color = c("#ff9999","#ff0000"),
x = "Fold.Enrichment", xlabel = "Fold Enrichment")
}
