% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplePlot.R
\name{David_dotplot}
\alias{David_dotplot}
\title{David富集结果绘图 气泡图}
\usage{
David_dotplot(
  df,
  term.prefix = c("(^GO:\\\\d+~)|(^\\\\w+\\\\d+[:~])"),
  fill.color = c("blue", "red"),
  show.top = 10,
  x = "Fold.Enrichment",
  pt.size = "Count",
  xlabel = "Fold Enrichment",
  size.label = "Gene number",
  arrange.by.x = FALSE,
  decresing = TRUE
)
}
\arguments{
\item{df}{David富集结果, A data.frame}

\item{term.prefix}{Term列中的前缀，会被去除}

\item{fill.color}{柱子的颜色,一个向量，分别对应最低值和最高值}

\item{show.top}{要展示的Term数目}

\item{x}{X轴的映射}

\item{pt.size}{点大小的映射}

\item{xlabel}{X轴的标签}

\item{size.label}{Legend点大小的label}

\item{arrange.by.x}{Term是否按照X轴的值进行排序，默认是按照PValue排序的}

\item{decresing}{排序方式默认为降序}
}
\value{
a ggplot2 plot
}
\description{
David富集结果绘图 气泡图
}
\examples{
requireNamespace("dplyr")
requireNamespace("ggplot2")
df <- read.delim(system.file("extdata", "David_outputs_KEGG.txt",
package = "ZhangRtools"), stringsAsFactors = FALSE)
David_dotplot(df)
David_dotplot(df, arrange.by.x = TRUE)
}
