% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameRelated.R
\name{Expand_df}
\alias{Expand_df}
\title{对某列进行字符串切割，复制其他所有行【扩展行】}
\usage{
Expand_df(df, id, splitby = "/")
}
\arguments{
\item{df}{A data.frame}

\item{id}{用于字符换切割的列}

\item{splitby}{字符串切割时的分隔符}
}
\value{
A data.frame with more rows
}
\description{
以字符串转表达式方式，实现某列字符串数据切割和行扩展。只能基于一列进行字符串切割。
}
\examples{
#NULL
}
