% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseqRelated.R
\name{FeaturePlot_Single}
\alias{FeaturePlot_Single}
\title{对单个基因绘制splited feature plot，支持使用common legend}
\usage{
FeaturePlot_Single(
  obj,
  feature,
  SplitBy,
  PointColor = c("#638ED0", "#ffff33", "#ff3300"),
  ncol = NULL,
  remove_axes = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object}

\item{feature}{a single feature}

\item{SplitBy}{group cells by}

\item{PointColor}{colors for low, middle, high expression}

\item{ncol}{number of columns to show plots}

\item{remove_axes}{是否需要删除xy坐标轴}

\item{...}{}
}
\description{
对单个基因绘制splited feature plot，支持使用common legend
}
\examples{
# FeaturePlot_Single(obj = cds, feature = "Lgr5", SplitBy = 'class', ncol = 4, remvove_axes = TRUE)
}
