% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameRelated.R
\name{mapping_join}
\alias{mapping_join}
\title{基于多个关键词合并两个数据}
\usage{
mapping_join(
  inputDF,
  db = ADataframe,
  by.input = NULL,
  by.db = NULL,
  by.input.2 = NULL,
  by.db.2 = NULL,
  by.input.3 = NULL,
  by.db.3 = NULL,
  label = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{inputDF}{原数据框}

\item{db}{要被加进来的数据}

\item{by.input}{原数据框里的第一识别列（也是要被更新的列）}

\item{by.db}{新数据框里的第一识别列的对应列}

\item{by.input.2}{原数据框里的第二识别列，当第一识别列不能被识别时才启用，可以为NULL}

\item{by.db.2}{新数据框里的第二识别列的对应列，可以为NULL}

\item{by.input.3}{原数据框里的第三识别列，当第一和第二识别列均不能被识别时才启用，可以为NULL}

\item{by.db.3}{新数据框里的第三识别列的对应列，可以为NULL}

\item{label}{默认为TRUE，添加label列，表征被第几对关键词所识别}

\item{verbose}{是否输出被第二、三个关键词匹配的条目的详细信息，用于人工校验准确性。}
}
\value{
A list，包含两个数据框，第一个matched：合并后的数据框，第二个lost，不能被识别的行.可能的问题：1. 如果两个ensembl id对应同一个HGNC，那么第二轮
用HGNC匹配时，可能仅能匹配到一个Ensebmbl id。同理第三轮匹配也是。
}
\description{
与mapping_update的区别，mapping_join不会修改原来的数据，只是在原来数据的基础上，添加新数据到新的列中。
依据多对关键词，合并两个数据。返回1个list，包含两个数据框，第一个matched：合并后的数据框，第二个lost，不能被识别的行。
注意：该函数的第一个关键词，即by.input列，应为绝大多数能被匹配上的。关键词可以重复使用，每次的关键词对不同即可。另外，该函数在后两轮匹配时，会输出重复匹配的条目。
关键词匹配是有优先顺序的，第一次被匹配上了，后续就不会再去做匹配。
}
\examples{
#NULL
}
