% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameRelated.R
\name{mapping_update}
\alias{mapping_update}
\title{基于多个关键词更新数据框}
\usage{
mapping_update(
  inputDF,
  db = ADataframe,
  by.input = NULL,
  by.db = NULL,
  by.input.2 = NULL,
  by.db.2 = NULL,
  by.input.3 = NULL,
  by.db.3 = NULL,
  label = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{inputDF}{要被更新的数据框}

\item{db}{数据库}

\item{by.input}{数据框里的第一识别列（也是要被更新的列）}

\item{by.db}{数据库里的第一识别列的对应列}

\item{by.input.2}{数据框里的第二识别列，当第一识别列不能被识别时才启用（不会被更新），可以为NULL}

\item{by.db.2}{数据库里的第二识别列的对应列，可以为NULL}

\item{by.input.3}{数据框里的第三识别列，当第一和第二识别列均不能被识别时才启用（不会被更新），可以为NULL}

\item{by.db.3}{数据库里的第三识别列的对应列，可以为NULL}

\item{label}{默认为TRUE，添加label列，表征被第几个关键词所识别}

\item{verbose}{是否输出被第二、三个关键词匹配的条目的详细信息，用于人工校验准确性。}
}
\value{
A list，包含两个数据框，第一个matched：更新后的数据框，第二个lost，不能被识别的行.可能的问题：1. 如果两个ensembl id对应同一个HGNC，那么第二轮
用HGNC匹配时，可能仅能匹配到一个Ensebmbl id。同理第三轮匹配也是。
}
\description{
依据数据库信息，和至多3种ID，更新数据。返回1个list，包含两个数据框，第一个matched：更新后的数据框，第二个lost，不能被识别的行。
注意：该函数值更新了第一个关键词，即by.input列。另外，该函数在后两轮匹配时，会输出重复匹配的条目。
注意：BiologyDB包依赖此函数，勿轻易修改
}
\examples{
#NULL
}
