% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchModifyFiles.R
\name{ModifyFile}
\alias{ModifyFile}
\title{修改单个文件里的内容}
\usage{
ModifyFile(
  file,
  LineMatchKeyWords,
  LineMatch.ignore.case = FALSE,
  WordOld,
  wordold.Matchfixed = TRUE,
  WordNew,
  Replace = FALSE,
  SaveOld = TRUE,
  silence = TRUE,
  returnSummary = FALSE
)
}
\arguments{
\item{file}{要被修改的文件,一个路径字符串}

\item{LineMatchKeyWords}{用于寻找要被修改的行,一个关键词向量}

\item{LineMatch.ignore.case}{依据关键词寻找要被修改的行时,是否需要忽略行
关键词的大小写,一个逻辑值}

\item{WordOld}{要被替换的词,一个字符串}

\item{wordold.Matchfixed}{匹配要被替换的词时,是否需要设置为不使用正则表达式,
一个逻辑值}

\item{WordNew}{新词,一个字符串}

\item{Replace}{是否要执行替换,一个逻辑值}

\item{SaveOld}{是否要保存旧的文件,一个逻辑值}

\item{silence}{是否屏蔽信息输出,一个逻辑值}

\item{returnSummary}{是否输出替换统计结果,一个逻辑值,主要用于批量替换}
}
\value{
A data.frame of 匹配和替换信息 or Nothing
}
\description{
以逐行读取方式,对文件里的内容进行内容匹配和替换。
}
\examples{
# Not Run
# ModifyFile(file = "./scRNAseq/Fly/Ovary/Fly-Ovary-Jevitt-PlosBiology-2020/
# Parameters.R",
# LineMatchKeyWords = c("SplitBy.levels.max","15"),
# LineMatch.ignore.case = FALSE,
# WordOld = "SplitBy.levels.max <- 15",
# wordold.Matchfixed = TRUE,
# WordNew = "SplitBy.levels.max <- 50",
# Replace = TRUE,
# SaveOld = TRUE,
# silence = FALSE,
# returnSummary = TRUE
# )

}
