% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseqRelated.R
\name{top_genes}
\alias{top_genes}
\title{各个细胞内的高表达基因}
\usage{
top_genes(SeuratObj, expr.cut = 0.01)
}
\arguments{
\item{SeuratObj}{Seurat对象}

\item{expr.cut}{针对UMI counts比例所设定的cut off，用于定义高表达的基因。}
}
\value{
A data.frame
}
\description{
对Seurat对象里的各个单细胞，分别统计top表达的基因，即基因的reads比例大于所设定的阈值expt.cut。
}
\examples{
# top_genes(SeuratObj = obj, expr.cut = 0.01)
}
