% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameRelated.R
\name{update_IDs}
\alias{update_IDs}
\title{基于数据库，对ID向量进行转换，返回更新后的向量}
\usage{
update_IDs(
  old,
  db = ADataframe,
  from = NULL,
  to = NULL,
  split = NULL,
  fixed = TRUE,
  collapse = ";"
)
}
\arguments{
\item{old}{要被转换的id向量，元素可以是多个id名的组合，用| ；或其它符号隔开。}

\item{db}{注释数据库，包含旧名和新名}

\item{from}{注释数据库中对应原来ID的列名}

\item{to}{注释数据库中新ID的列名}

\item{split}{如果ID向量为多id组合的，需要指定分割符号，默认为NULL，即为单个ID，无需分割}

\item{fixed}{默认TRUE，strsplit函数的参数}
}
\value{
转换后的向量
}
\description{
NA值或空字符串会返回NA值，ID向量里的单个元素可以是多个ID的组合。
注意：BiologyDB包依赖此函数，勿轻易修改
}
\examples{
#NULL
}
