% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplePlot.R
\name{Volcano_plot_1}
\alias{Volcano_plot_1}
\title{火山图 风格一}
\usage{
Volcano_plot_1(
  plot.data,
  markers,
  pvalue.cutoff = 0.05,
  lfc.cutoff = 1,
  x.lim = c(-10, 10)
)
}
\arguments{
\item{markers}{genes in gene column to be labeled}

\item{data}{a data.frame with columns: gene, pval, lfc, 默认lfc以2为底}
}
\value{
a plot
}
\description{
火山图 风格一
}
\examples{
# null
}
